refactor: moved the docs_generation folder to docs IHS-201
