--- doc_generation/content_gen_methods/command/__init__.py
@@ -1,11 +1,0 @@
-from __future__ import annotations
-
-from .command import ACommand
-from .typer_command import ATyperCommand, TyperGroupCommand, TyperSingleCommand
-
-__all__ = [
-    "ACommand",
-    "ATyperCommand",
-    "TyperGroupCommand",
-    "TyperSingleCommand",
-]
