+++ docs/docs_generation/content_gen_methods/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from doc_generation.content_gen_methods.command.command import ACommand
-from doc_generation.content_gen_methods.command.typer_command import TyperGroupCommand, TyperSingleCommand
-from doc_generation.content_gen_methods.mdx.mdx_code_doc import MdxCodeDocumentation
+from docs.docs_generation.content_gen_methods.command.command import ACommand
+from docs.docs_generation.content_gen_methods.command.typer_command import TyperGroupCommand, TyperSingleCommand
+from docs.docs_generation.content_gen_methods.mdx.mdx_code_doc import MdxCodeDocumentation
 
 from .base import ADocContentGenMethod
 from .command_output_method import CommandOutputDocContentGenMethod
