+++ docs/docs_generation/content_gen_methods/command/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .command import ACommand
+
+__all__ = [
+    "ACommand",
+]
