+++ docs/docs_generation/content_gen_methods/command_output_method.py
