+++ docs/docs_generation/content_gen_methods/mdx/__init__.py
@@ -6,11 +6,3 @@ __all__ = [
     "MdxCodeDocumentation",
     "MdxFile",
 ]
-from __future__ import annotations
-
-from .mdx_code_doc import MdxCodeDocumentation, MdxFile
-
-__all__ = [
-    "MdxCodeDocumentation",
-    "MdxFile",
-]
