+++ tests/unit/doc_generation/content_gen_methods/mdx/test_mdx_code_doc.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from invoke import Result
 from invoke.context import MockContext
 
-from doc_generation.content_gen_methods import (
+from docs.docs_generation.content_gen_methods import (
     MdxCodeDocumentation,
 )
